fix(react-payment-brand-icons): move IconManifestEntry to src/ so generated manifest stays in lib rootDir
Also include src/types.ts in tsconfig.node.json (scripts/types.ts now re-exports from
src/) and add DOM lib to tsconfig.test.json (required for the IconShell test's DOM
queries; the test config previously inherited lib:ESNext only, predating any React
component in this package).
