+++ packages/react-payment-brand-icons/src/types.ts
@@ -1,0 +1,6 @@
+export type IconManifestEntry = {
+    slug: string;
+    componentName: string;
+    title: string;
+    aliases: readonly string[];
+};
