+++ packages/react-payment-brand-icons/tsconfig.node.json
@@ -1,5 +1,5 @@
 {
     "$schema": "https://json.schemastore.org/tsconfig",
     "extends": "../../tsconfig.node.json",
-    "include": ["vitest.config.ts", "scripts/**/*.ts", "icons.ts"]
+    "include": ["vitest.config.ts", "scripts/**/*.ts", "icons.ts", "src/types.ts"]
 }
