+++ packages/react-payment-brand-icons/tsconfig.test.json
@@ -3,6 +3,7 @@
     "extends": ["./tsconfig.json", "../../tsconfig.test.json"],
     "compilerOptions": {
         "types": ["node", "vitest/globals"],
+        "lib": ["ESNext", "DOM", "DOM.Iterable"],
         "rootDir": "${configDir}",
         "composite": false,
         "declaration": false,
